Refactor Service Gateway logic; remove unused files and functions, enhance diff tracker integration, and improve error handling
