--- pkg/provider/azure_servicegateway.go
@@ -1,27 +1,0 @@
-/*
-Copyright 2026 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package provider
-
-import (
-	"strings"
-
-	v1 "k8s.io/api/core/v1"
-)
-
-func getServiceUID(service *v1.Service) string {
-	return strings.ToLower(string(service.UID))
-}
