+++ pkg/provider/azure_standard.go
@@ -44,6 +44,7 @@ import (
 	"sigs.k8s.io/cloud-provider-azure/pkg/consts"
 	"sigs.k8s.io/cloud-provider-azure/pkg/log"
 	"sigs.k8s.io/cloud-provider-azure/pkg/metrics"
+	"sigs.k8s.io/cloud-provider-azure/pkg/provider/servicegateway/difftracker"
 	vmutil "sigs.k8s.io/cloud-provider-azure/pkg/util/vm"
 )
 
@@ -319,7 +320,7 @@ func (az *Cloud) getRulePrefix(service *v1.Service) string {
 
 func (az *Cloud) getPublicIPName(clusterName string, service *v1.Service, isIPv6 bool) (string, error) {
 	if az.ServiceGatewayEnabled {
-		return fmt.Sprintf("%s-pip", getServiceUID(service)), nil
+		return difftracker.PublicIPName(difftracker.ServiceUID(service)), nil
 	}
 
 	logger := log.Background().WithName("getPublicIPName")
